feat: add express server bootstrap
