+++ package.json
@@ -3,12 +3,12 @@
   "version": "0.13.3",
   "author": "Francisco Calderón <negrunch@grunch.dev>",
   "description": "P2P lightning network telegram bot",
-  "main": "dist/app.js",
+  "main": "dist/server.js",
   "scripts": {
     "prestart": "npx tsc",
-    "start": "node ./dist/app",
+    "start": "node ./dist/server",
     "predev": "npx tsc",
-    "dev": "nodemon ./dist/app",
+    "dev": "nodemon ./dist/server",
     "lint": "eslint . --ext .ts",
     "format": "prettier --write '**/*.{js,ts}'",
     "pretest": "tsc -p tsconfig.test.json",
@@ -22,6 +22,7 @@
     "canvas": "^3.0.0",
     "crypto": "^1.0.1",
     "dotenv": "^10.0.0",
+    "express": "^5.1.0",
     "invoices": "2.0.6",
     "lightning": "10.25.0",
     "mongoose": "6.13.6",
@@ -41,6 +42,7 @@
     "network"
   ],
   "devDependencies": {
+    "@types/express": "^5.0.3",
     "@types/i18n": "^0.13.12",
     "@types/mocha": "^10.0.9",
     "@types/node": "^20.16.15",
